requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
black==21.7b0
certifi==2019.11.28
cffi==1.14.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==2.8
dataclasses==0.8
decorator==4.4.2
future==0.18.2
idna==2.10
importlib-metadata==4.7.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
pathspec==0.9.0
protobuf==3.14.0
pycparser==2.20
python-dotenv==0.15.0
python-telegram-bot==12.4.2
pytz==2019.3
regex==2021.8.28
requests==2.25.1
six==1.14.0
tomli==1.2.1
tornado==6.0.4
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.3
zipp==3.5.0
